feat: move core settings to a sphinx extension
Move all the other settings beside doxygen to a sphinx extension.
Backwards compatibility is kept by enabling the extension by default
in `rocm_docs.setup`. With the extension make most settings overridable
from `conf.py`.

The settings are described declaratively based on what needs to happen
when the user also configured it (see _ConfigDefault, _ConfigExtend, etc
classes).

using rocm-docs-core this way is as easy as having
```python
extensions = ["rocm_docs"]
```
in `conf.py` if doxygen is not required.

The doxygen parts are not yet fully moved to an extension, but the
`rocm_docs.doxygen` module is added (to make the rocm_docs name free
for the core extension.). Using these still requires the "old" approach.