feat: Add support for theme "flavors"
Most parts of rocm-docs-core are generic and don't assume the consuming
project is part of the ROCm software suite, but there are some branding
related settings that do but consuming projects wouldn't want to duplicate
(Like logos / headers & footers).

rocm-docs-core wants to support not just ROCm documentation, so these
elements should be configurable but still centralized. The idea is that
the ecosystem specific parts are organized to be "flavor"s of rocm-docs-core.
Each flavor customizes these branding settings, and consumers can chose
the flavor they want.
Currently the only supported flavor is the default ROCm flavor, so this
should result in no changes for now.

Currently flavors have to be defined in this repository (rocm-docs-core),
but it might be a good idea to allow installed python packages to
provide additional flavors.