feat(doxygen,projects): Make doxygen tagfile available
Allow specifying a doxygen html directory in projects.yaml. From
this location copy a file named "tagfile.xml" to the output.
This file is created by doxygen (based on the setting `GENERATE_TAGFILE`)
and describes the symbols in the doxygen documentation.

This will be used to allow linking to doxygen symbols from sphinx
pages in- and cross-project.

The setting in projects.yaml can be overriden with the configuration
option `doxygen_html`. This is to allow initial setup (where rocm-docs-core)
doesn't (yet) have the setting in projects.yaml or migration to a different
doxygen path.

Without this option `projects.yaml` would always have to be modified first,
before the project can be changed, but it would point to an invalid
location until the project is rebuilt.