feat(doxygen.py): automatic setup of doxylink
**If** doxylink is enabled by the project
(by adding `sphinxcontrib.doxylink` to `extensions`) and `doxygen_html`
is set, then set the settings of doxylink automatically.
The role name is set to "doxygen", therefore references to the doxygen
documentation in the current project can be typed as
`{doxygen}\`entity-in-doxygen-doc\`` which could be a file, functions,
namespace etc.

This is only covers references within the same project, outside
references will be done later.